DYGYM-16
* 디자인 변경 및 삭제되지 않는 운동 버그 수정

+++ pages/management/index.jsx
@@ -82,6 +82,11 @@ export default function Management() {
   };
 
   const deleteCategory = async (e, id) => {
+    const findCategory = data.find((d) => d.id === id);
+    if (findCategory && findCategory.count > 0) {
+      alert("저장된 운동을 모두 지워야 분류를 지울 수 있습니다");
+      return;
+    }
     await axios.delete(`/category?id=${id}`);
     mutate("/get/category");
     setModal({
